moved copyCaseNumber to lib/utils and renamed formatCaseNumber
Jira ticket: CAMS-428
