+++ user-interface/src/data-verification/transfer/PendingTransferOrder.test.tsx
@@ -9,7 +9,7 @@ import {
 import { BrowserRouter } from 'react-router-dom';
 import { MockData } from '@common/cams/test-utilities/mock-data';
 import { UswdsAlertStyle } from '@/lib/components/uswds/Alert';
-import { getCaseNumber } from '@/lib/utils/formatCaseNumber';
+import { getCaseNumber } from '@/lib/utils/caseNumber';
 import Api2 from '@/lib/models/api2';
 
 const fromCaseSummary = MockData.getCaseSummary();
